@@ -5,7 +5,7 @@
   "scripts": {
     "build": "npm run-script lint && cross-env NODE_ENV=production webpack --config webpack.prod.js",
     "build:stats": "npm run-script lint && cross-env NODE_ENV=production:stats webpack --config webpack.prod.js --profile --json > stats.json",
-    "start:dev": "cross-env NODE_ENV=development webpack serve --hot --config webpack.dev.js",
+    "start:dev": "cross-env NODE_ENV=development NODE_OPTIONS=\"--loader ts-node --experimental-modules\" npx webpack serve --hot -c webpack.dev.ts",
     "lint": "eslint --fix ./src",
     "deploy": "npm run build && firebase deploy"
   },
@@ -14,10 +14,10 @@
     "./src/utils/prototypes/*.ts"
   ],
   "dependencies": {
-    "@firebase/app": "^0.7.17",
-    "@firebase/auth": "^0.19.9",
-    "@firebase/firestore": "^3.4.5",
-    "@firebase/util": "^1.4.3",
+    "@firebase/app": "^0.7.20",
+    "@firebase/auth": "^0.19.11",
+    "@firebase/firestore": "^3.4.7",
+    "@firebase/util": "^1.5.1",
     "@material/material-color-utilities": "^0.1.1",
     "@material/mwc-button": "^0.25.3",
     "@material/mwc-dialog": "^0.25.3",
@@ -28,27 +28,31 @@
     "@material/mwc-menu": "^0.25.3",
     "@material/mwc-slider": "^0.25.3",
     "@material/mwc-textfield": "^0.25.3",
-    "@sentry/browser": "^6.18.2",
-    "@sentry/tracing": "^6.18.2",
-    "lit": "^2.2.0",
+    "@sentry/browser": "^6.19.2",
+    "@sentry/tracing": "^6.19.2",
+    "lit": "^2.2.1",
     "md5": "^2.3.0"
   },
   "devDependencies": {
     "@sentry/webpack-plugin": "^1.18.8",
     "@types/css-minimizer-webpack-plugin": "^3.2.0",
     "@types/md5": "^2.3.2",
-    "@types/node": "^17.0.21",
+    "@types/node": "^17.0.23",
     "@types/terser-webpack-plugin": "^5.2.0",
     "@types/webpack": "^5.28.0",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
+    "@types/webpack-dev-server": "^4.7.1",
+    "@typescript-eslint/eslint-plugin": "^5.16.0",
+    "@typescript-eslint/parser": "^5.16.0",
     "@webpack-cli/init": "^1.1.3",
+    "babel-loader": "^8.2.4",
+    "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
+    "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^3.4.1",
     "esbuild-loader": "^2.18.0",
-    "eslint": "^8.10.0",
+    "eslint": "^8.12.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-import-resolver-typescript": "^2.7.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-lit": "^1.6.1",
     "eslint-plugin-wc": "^1.3.2",
@@ -57,15 +61,18 @@
     "inspectpack": "^4.7.1",
     "lit-css-loader": "^1.2.2",
     "minify-html-literals-loader": "^1.1.1",
+    "nodemon": "^2.0.15",
     "sass": "^1.49.9",
     "sass-loader": "^12.6.0",
     "simple-progress-webpack-plugin": "^2.0.0",
+    "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
-    "ts-node": "^10.5.0",
+    "ts-loader": "^9.2.8",
+    "ts-node": "^10.7.0",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "typescript": "^4.6.2",
+    "typescript": "^4.6.3",
     "typescript-plugin-css-modules": "^3.4.0",
-    "webpack": "^5.69.1",
+    "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4",
     "webpack-merge": "^5.8.0"
