@@ -1,4 +1,4 @@
-import { CSSResultGroup, html, LitElement, PropertyValues, TemplateResult } from "lit";
+import { CSSResultGroup, html, LitElement, PropertyValues, TemplateResult, unsafeCSS } from "lit";
 import { property, query } from "lit/decorators.js";
 import { componentStyles } from "~src/global";
 import { defineComponent } from "~utils/components";
@@ -212,6 +212,6 @@ export class ProgressLine extends LitElement {
     }
 
     static get styles(): CSSResultGroup {
-        return [...componentStyles, scopedStyles as never];
+        return [...componentStyles, unsafeCSS(scopedStyles)];
     }
 }
