@@ -1,4 +1,4 @@
-import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
+import { css, CSSResultGroup, html, LitElement, TemplateResult, unsafeCSS } from "lit";
 import { customElement } from "lit/decorators.js";
 import { pageStyles } from "~src/global";
 import { getAuth } from "@firebase/auth";
@@ -23,7 +23,7 @@ export default class LandingPage extends LitElement {
     static get styles(): CSSResultGroup {
         // Styles can either be in this file (only css)
         // or imported from another file (scss in this case)
-        return [...pageStyles, scopedStyles as never, css`
+        return [...pageStyles, unsafeCSS(scopedStyles), css`
           // More styles here
         `];
     }
